requirements.txt
django
south
djangorestframework
django-extensions
werkzeug
django-cors-headers
git+https://github.com/omab/python-social-auth.git
gunicorn
django-compressor
django-allauth
django-debug-toolbar